# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Approval System API
  version: 1.0.0
  description: 基于当前代码实现整理的接口文档（Spring Boot + Flowable）
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/forms/definitions:
    post:
      tags: [Forms]
      summary: 创建表单定义（需 DESIGNER）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormDefinitionRequest'
            example:
              userId: 2001
              formKey: travel_request
              formName: 差旅申请表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormDefinition'
              example:
                id: 1
                formName: 差旅申请表
                formKey: travel_request
                status: 1
        '403':
          description: 权限不足
  /api/forms/versions:
    post:
      tags: [Forms]
      summary: 创建表单版本（需 DESIGNER）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormVersionRequest'
            example:
              userId: 2001
              formId: 1
              schemaJson: "{\"title\":\"差旅申请\",\"type\":\"object\"}"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormVersion'
              example:
                id: 3
                formId: 1
                version: 3
                schemaJson: "{\"title\":\"差旅申请\",\"type\":\"object\"}"
        '403':
          description: 权限不足
  /api/forms/fields:
    post:
      tags: [Forms]
      summary: 替换表单字段（需 DESIGNER）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceFieldsRequest'
            example:
              userId: 2001
              formVersionId: 3
              fields:
                - fieldKey: destination
                  fieldType: string
                  label: 出差地点
                  required: true
                - fieldKey: budget
                  fieldType: number
                  label: 预算金额
                  required: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Fields updated
        '403':
          description: 权限不足
    get:
      tags: [Forms]
      summary: 获取表单字段
      parameters:
        - name: formVersionId
          in: query
          required: true
          description: 表单版本ID
          schema:
            type: integer
            format: int64
            example: 3
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FormField'
              example:
                - id: 11
                  formVersionId: 3
                  fieldKey: destination
                  fieldType: string
                  label: 出差地点
                  required: 1
                  visibleRule: null
                  validateRule: null
                  optionsJson: null
  /api/forms/versions/latest:
    get:
      tags: [Forms]
      summary: 获取指定表单Key最新版本
      parameters:
        - name: formKey
          in: query
          required: true
          description: 表单唯一键
          schema:
            type: string
            example: travel_request
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormVersion'
              example:
                id: 3
                formId: 1
                version: 3
                schemaJson: "{\"title\":\"差旅申请\",\"type\":\"object\"}"
  /api/forms/instances:
    post:
      tags: [Forms]
      summary: 创建表单实例（需 EMPLOYEE）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormInstanceRequest'
            example:
              userId: 1001
              formVersionId: 3
              businessKey: BR-20260224-0001
              data:
                destination: 上海
                budget: 3200
                days: 3
                reason: 客户拜访
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormInstance'
              example:
                id: 88
                formVersionId: 3
                businessKey: BR-20260224-0001
                dataJson: "{\"destination\":\"上海\",\"budget\":3200,\"days\":3}"
        '400':
          description: 参数或校验失败
        '403':
          description: 权限不足
  /api/forms/validate:
    post:
      tags: [Forms]
      summary: 校验表单数据（需 EMPLOYEE）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateFormInstanceRequest'
            example:
              userId: 1001
              formVersionId: 3
              data:
                destination: 上海
                budget: 3200
                days: 3
      responses:
        '200':
          description: 校验通过
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Validation passed
        '400':
          description: 校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormValidationErrorResponse'
              example:
                error: Form validation failed
                errors:
                  - field: budget
                    code: required
                    message: budget is required
        '403':
          description: 权限不足
...
components:
...
  schemas:
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: true
        message:
          type: string
          description: 操作结果说明
          example: Task completed
...
    FormDefinition:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 表单定义ID
          example: 1
        formName:
          type: string
          description: 表单名称
          example: 差旅申请表
        formKey:
          type: string
          description: 表单唯一键
          example: travel_request
        status:
          type: integer
          description: 状态（0禁用/1启用）
          example: 1
    FormVersion:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 表单版本ID
          example: 3
        formId:
          type: integer
          format: int64
          description: 表单定义ID
          example: 1
        version:
          type: integer
          description: 版本号
          example: 3
        schemaJson:
          type: string
          description: 表单Schema JSON
          example: "{\"title\":\"差旅申请\",\"type\":\"object\"}"
    FormField:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 字段ID
          example: 11
        formVersionId:
          type: integer
          format: int64
          description: 表单版本ID
          example: 3
        fieldKey:
          type: string
          description: 字段标识
          example: destination
        fieldType:
          type: string
          description: 字段类型
          example: string
        label:
          type: string
          description: 字段名称
          example: 出差地点
        required:
          type: integer
          description: 是否必填（0否/1是）
          example: 1
        visibleRule:
          type: string
          nullable: true
          description: 可见性规则
          example: budget > 0
        validateRule:
          type: string
          nullable: true
          description: 校验规则
          example: max:10000
        optionsJson:
          type: string
          nullable: true
          description: 选项配置JSON
          example: "[\"上海\",\"北京\"]"
    FormInstance:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 表单实例ID
          example: 88
        formVersionId:
          type: integer
          format: int64
          description: 表单版本ID
          example: 3
        businessKey:
          type: string
          description: 业务主键
          example: BR-20260224-0001
        dataJson:
          type: string
          description: 表单数据JSON
          example: "{\"destination\":\"上海\",\"budget\":3200}"
    CreateFormDefinitionRequest:
      type: object
      required: [userId, formKey, formName]
      properties:
        userId:
          type: integer
          format: int64
          description: 操作人用户ID（需 DESIGNER）
          example: 2001
        formKey:
          type: string
          description: 表单唯一键
          example: travel_request
        formName:
          type: string
          description: 表单名称
          example: 差旅申请表
    CreateFormVersionRequest:
      type: object
      required: [userId, formId, schemaJson]
      properties:
        userId:
          type: integer
          format: int64
          description: 操作人用户ID（需 DESIGNER）
          example: 2001
        formId:
          type: integer
          format: int64
          description: 表单定义ID
          example: 1
        schemaJson:
          type: string
          description: 表单Schema定义
          example: "{\"title\":\"差旅申请\",\"type\":\"object\"}"
    ReplaceFieldsRequest:
      type: object
      required: [userId, formVersionId, fields]
      properties:
        userId:
          type: integer
          format: int64
          description: 操作人用户ID（需 DESIGNER）
          example: 2001
        formVersionId:
          type: integer
          format: int64
          description: 表单版本ID
          example: 3
        fields:
          type: array
          description: 字段定义列表
          example:
            - fieldKey: destination
              fieldType: string
              label: 出差地点
              required: true
          items:
            $ref: '#/components/schemas/FormFieldRequest'
    FormFieldRequest:
      type: object
      properties:
        fieldKey:
          type: string
          description: 字段标识
          example: destination
        fieldType:
          type: string
          description: 字段类型（string/number/date/select/table）
          example: string
        label:
          type: string
          description: 字段标题
          example: 出差地点
        required:
          type: boolean
          description: 是否必填
          example: true
        visibleRule:
          type: string
          nullable: true
          description: 可见性规则表达式
          example: budget > 0
        validateRule:
          type: string
          nullable: true
          description: 校验规则表达式
          example: max:10000
        optionsJson:
          type: string
          nullable: true
          description: 选项JSON（select/table等字段使用）
          example: "[\"上海\",\"北京\"]"
    CreateFormInstanceRequest:
      type: object
      required: [userId, formVersionId, businessKey, data]
      properties:
        userId:
          type: integer
          format: int64
          description: 操作人用户ID（需 EMPLOYEE）
          example: 1001
        formVersionId:
          type: integer
          format: int64
          description: 表单版本ID
          example: 3
        businessKey:
          type: string
          description: 业务主键
          example: BR-20260224-0001
        data:
          type: object
          additionalProperties: true
          description: 表单数据键值对
          example:
            destination: 上海
            budget: 3200
            days: 3
    ValidateFormInstanceRequest:
      type: object
      required: [userId, formVersionId, data]
      properties:
        userId:
          type: integer
          format: int64
          description: 操作人用户ID（需 EMPLOYEE）
          example: 1001
        formVersionId:
          type: integer
          format: int64
          description: 表单版本ID
          example: 3
        data:
          type: object
          additionalProperties: true
          description: 待校验的表单数据
          example:
            destination: 上海
            budget: 3200
    FormValidationError:
      type: object
      properties:
        field:
          type: string
          description: 错误字段
          example: budget
        code:
          type: string
          description: 错误码
          example: required
        message:
          type: string
          description: 错误信息
          example: budget is required
    FormValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 总体错误描述
          example: Form validation failed
        errors:
          type: array
          description: 具体字段错误列表
          items:
            $ref: '#/components/schemas/FormValidationError'
...